+++ include/boe/adventurers.hpp
@@ -5,3 +5,4 @@
 bool cave_lore_present(const Adventurers & adven);
 bool woodsman_present(const Adventurers& adven);
 short mage_lore_total(const Adventurers& adven);
+void heal_party(Adventurers& adventurers, short amt);
