+++ include/boe/pc.hpp
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "boe/game_file_structures.hpp"
+
+void heal_pc(pc_record_type& pc, short amt);
