+++ src/Windows Code Release 3/party.cpp
@@ -757,27 +757,6 @@ Boolean create_pc(short spot,short parent_num)
 	return TRUE;
 }
 
-void heal_pc(pc_record_type & pc,short amt)
-{
-	if (pc.cur_health > pc.max_health)
-		return;
-	if (pc.main_status != 1)
-		return;
-	pc.cur_health += amt;
-	if (pc.cur_health > pc.max_health)
-		pc.cur_health = pc.max_health;
-
-}
-
-void heal_party(Adventurers& adventurers, short amt)
-{
-	short i;
-	
-	for (i = 0; i < 6; i++)
-		if (adventurers[i].main_status == 1)
-			heal_pc(adventurers[i],amt);
-}
-
 void cure_pc(short pc_num,short amt)
 {
 	if (adven[pc_num].main_status != 1)
