+++ src/Windows Code Release 3/party.h
@@ -1,3 +1,4 @@
+#include "boe/pc.hpp"
 #include "boe/adventurers.hpp"
 
 void init_party(short mode);
@@ -7,8 +8,6 @@ pc_record_type return_dummy_pc();
 pc_record_type create_debug_pc(short num);
 pc_record_type create_prefab_pc(short num);
 Boolean create_pc(short spot,short parent_num);
-void heal_pc(pc_record_type& pc,short amt);
-void heal_party(Adventurers& adventurers, short amt);
 void cure_pc(short pc_num,short amt);
 void cure_party(short amt);
 void curse_pc(short which_pc,short how_much);
